+++ Src/HurryUp_Agent/CUtils.h
@@ -15,7 +15,7 @@ struct sockaddr_ll
 void SetLogger(std::tstring _name, DWORD _inputOption);
 std::tstring ReadContent(const char* path);
 std::vector<std::tstring> Split(std::tstring input, std::tstring delimiter);
-int FindFileEndPosition(std::ifstream& file);
+long long int FindFileEndPosition(std::ifstream& file);
 std::string GetTimeStamp();
 void CheckDirectory(std::tstring _path);
 std::string GeneratorStringNumber();
